KMRL-/requirements.txt
# Core dependencies
sqlalchemy>=1.4.0
python-magic>=0.4.24
python-magic-bin>=0.4.14  # For Windows

# PDF processing
pdfplumber>=0.7.0
pdf2image>=1.16.0

# Image processing and OCR
Pillow>=9.0.0
pytesseract>=0.3.10

# Office documents
python-docx>=0.8.11
pandas>=1.3.0
openpyxl>=3.0.9

# CAD files
ezdxf>=0.17.0

# Email processing
imapclient>=2.3.0
pyzmail36>=1.0.4

# File watching
watchdog>=2.1.0

# SharePoint (optional)
# Office365-REST-Python-Client>=2.3.11

# Utilities
requests>=2.25.0